+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.3.32] - 2019-01-16
+### Changed
+- UI and HTML changes for accessibility.
+### Fixed
+- Timing of initialization of Google autocomplete fields resulted in
+  JavaScript error in some circumstances.
+- OneDrive synchronization errors due to changes in the Microsoft Graph API.
+
 ## [0.3.31] - 2019-01-14
 ### Changed
 - Additional information added to data representation of question.